pyproject.toml
[project]
name = "resume-agent"
version = "0.1.0"
description = "简历增强 Agent - 基于 LangGraph 的智能简历优化服务"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.0",
    "deepagents>=0.1.0",
    "fastapi>=0.115.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.50",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "aiosqlite>=0.20.0",
    "openai>=1.50.0",
    "psycopg[binary,pool]>=3.2.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.1.0",
    "uvicorn>=0.32.0",
    "pymupdf>=1.26.7",
    "python-multipart>=0.0.22",
    "pillow>=12.1.0",
    "pymupdf4llm>=0.2.9",
    "bcrypt>=4.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.setuptools.packages.find]
where = ["apps/backend"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
check_untyped_defs = true
show_error_codes = true
pretty = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["apps/backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["backend", "config", "llm", "workflows", "infrastructure"]
